s1/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
kafka-python==2.0.2
faker==24.9.0
